requirements.txt
python-telegram-bot
psycopg
Flask
psycopg[binary]
sqlalchemy
pytz